hw2/requirements.txt
fastapi
uvicorn[standard]
pandas
matplotlib
scipy
sklearn
numpy
pyaml
requests
